+++ 约吧/Profile/Controllers/QYSelectionController.h
@@ -7,50 +7,21 @@
 //
 
 #import <UIKit/UIKit.h>
-#import "QYSelectModel.h"
-
-typedef NS_ENUM(NSUInteger, QYSelectionType) {
-    /** 职业 */
-    QYSelectionTypeOccupation,
-    
-    /** 来自 */
-    QYSelectionTypeHometown,
-    
-    /** 我的个性标签 */
-    QYSelectionTypePersonality,
-    
-    /** 我喜欢的运动 */
-    QYSelectionTypeSports,
-    
-    /** 我喜欢的音乐 */
-    QYSelectionTypeMusic,
-    
-    /** 我喜欢的食物 */
-    QYSelectionTypeFood,
-    
-    /** 我喜欢的电影 */
-    QYSelectionTypeMovies,
-    
-    /** 我喜欢的书和动漫 */
-    QYSelectionTypeLiterature,
-    
-    /** 我的旅行足迹 */
-    QYSelectionTypePlaces
-};
-
+#import "ProfileCommon.h"
 @class QYSelectionController, QYSelectModel;
 
 @protocol QYSelectionControllerDelegate <NSObject>
 
-- (void)selectionController:(QYSelectionController *)selectionController didSelectSelectModel:(QYSelectModel *)selectModel indexPath:(NSIndexPath *)indexPath;
+- (void)selectionController:(QYSelectionController *)selectionController didSelectSelectStrings:(NSArray *)selectStrings;
 
 @end
 
 @interface QYSelectionController : UITableViewController
 
-@property (nonatomic, strong) NSIndexPath *indexPath;
+@property (nonatomic, strong) NSString *selectedString;     //已经选中的描述
 
 @property (nonatomic, assign) QYSelectionType type;
+@property (nonatomic, assign) QYCreateTextType createTextType;
 @property (nonatomic, weak) id<QYSelectionControllerDelegate> delegate;
 
 @end
