+++ 约吧/Profile/Controllers/QYSubselectionController.h
@@ -7,7 +7,9 @@
 //
 
 #import <UIKit/UIKit.h>
-
+@class QYSelectModel;
+typedef void (^QYSelectedSubModel)(QYSelectModel *model);
 @interface QYSubselectionController : UITableViewController
-
+@property (nonatomic, strong) NSArray *subSelectionItems;
+@property (nonatomic, strong) QYSelectedSubModel selectedSubModel;
 @end
