+++ 约吧/Profile/Views/QYAutoHeightCell.xib
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="10116" systemVersion="15B42" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="10117" systemVersion="15D21" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES">
     <dependencies>
         <deployment identifier="iOS"/>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="10085"/>
@@ -28,7 +28,7 @@
                         <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
                     </view>
                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="我的个性标签" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="PPv-6M-7Da">
-                        <rect key="frame" x="45" y="8" width="84" height="27.5"/>
+                        <rect key="frame" x="45" y="14" width="84" height="15.5"/>
                         <constraints>
                             <constraint firstAttribute="height" constant="28" id="NUl-FP-eiW"/>
                         </constraints>
@@ -49,9 +49,9 @@
                     <constraint firstAttribute="bottom" secondItem="F86-S4-cwb" secondAttribute="bottom" constant="8" id="UO9-3i-Vzp"/>
                     <constraint firstItem="3cb-MR-Wou" firstAttribute="leading" secondItem="9CQ-jS-xZC" secondAttribute="leading" constant="15" id="Ywm-S8-ch0"/>
                     <constraint firstItem="3cb-MR-Wou" firstAttribute="centerY" secondItem="9CQ-jS-xZC" secondAttribute="centerY" id="efz-T1-nMX"/>
-                    <constraint firstAttribute="bottomMargin" secondItem="PPv-6M-7Da" secondAttribute="bottom" id="lYi-aa-O8i"/>
+                    <constraint firstItem="PPv-6M-7Da" firstAttribute="bottom" secondItem="9CQ-jS-xZC" secondAttribute="bottomMargin" constant="-6" id="lYi-aa-O8i"/>
                     <constraint firstItem="F86-S4-cwb" firstAttribute="leading" secondItem="3cb-MR-Wou" secondAttribute="trailing" constant="8" id="rXj-cy-ULw"/>
-                    <constraint firstItem="PPv-6M-7Da" firstAttribute="top" secondItem="9CQ-jS-xZC" secondAttribute="topMargin" id="tMt-t7-uUq"/>
+                    <constraint firstItem="PPv-6M-7Da" firstAttribute="top" secondItem="9CQ-jS-xZC" secondAttribute="topMargin" constant="6" id="tMt-t7-uUq"/>
                 </constraints>
             </tableViewCellContentView>
             <connections>
