+++ 约吧/Profile/Views/QYNormalHeightCell.xib
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="10116" systemVersion="15B42" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="10117" systemVersion="15D21" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES">
     <dependencies>
         <deployment identifier="iOS"/>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="10085"/>
+        <capability name="Constraints to layout margins" minToolsVersion="6.0"/>
     </dependencies>
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
@@ -21,7 +22,7 @@
                         <nil key="highlightedColor"/>
                     </label>
                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="添加职业信息" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="r7B-YJ-Cbu">
-                        <rect key="frame" x="79" y="13" width="84" height="17"/>
+                        <rect key="frame" x="108" y="13" width="84" height="17"/>
                         <fontDescription key="fontDescription" type="system" pointSize="14"/>
                         <color key="textColor" white="0.66666666666666663" alpha="1" colorSpace="calibratedWhite"/>
                         <nil key="highlightedColor"/>
@@ -29,8 +30,8 @@
                 </subviews>
                 <constraints>
                     <constraint firstItem="vic-vs-bIk" firstAttribute="centerY" secondItem="3FL-6s-RnE" secondAttribute="centerY" id="4gm-oQ-tJk"/>
+                    <constraint firstItem="r7B-YJ-Cbu" firstAttribute="leading" secondItem="3FL-6s-RnE" secondAttribute="leadingMargin" constant="100" id="Gxz-Ww-0cb"/>
                     <constraint firstItem="vic-vs-bIk" firstAttribute="leading" secondItem="3FL-6s-RnE" secondAttribute="leading" constant="15" id="VP3-ZT-n8s"/>
-                    <constraint firstItem="r7B-YJ-Cbu" firstAttribute="leading" secondItem="vic-vs-bIk" secondAttribute="trailing" constant="8" id="e1X-Q9-4AG"/>
                     <constraint firstItem="r7B-YJ-Cbu" firstAttribute="centerY" secondItem="vic-vs-bIk" secondAttribute="centerY" id="gRv-qS-W5Y"/>
                 </constraints>
             </tableViewCellContentView>
